requirements.txt
matplotlib==3.3.4
numpy==1.20.1
scipy==1.6.2
opencv-python==4.5.2.54
torch==1.9.0
torchvision==0.8.0
torchgeometry==0.1.2
scikit-image==0.18.1
Pillow==8.2.0